+++ packages/moneystream-extension/src/popup/components/util/Container.ts
@@ -1,8 +1,8 @@
 import { styled } from '@material-ui/core'
 
 export const Container = styled('div')(({ theme }) => ({
-  paddingRight: `${theme.spacing(2)}px`,
-  paddingLeft: `${theme.spacing(2)}px`,
+  paddingRight: `${theme.spacing(1)}px`,
+  paddingLeft: `${theme.spacing(1)}px`,
   marginRight: 'auto',
   marginLeft: 'auto',
   width: '100%',