+++ packages/moneystream-extension/src/popup/components/util/StatusTypography.ts
@@ -1,6 +1,6 @@
 import { styled, Typography } from '@material-ui/core'
-
+//${theme.spacing(1)}
 export const StatusTypography = styled(Typography)(({ theme }) => ({
-  paddingTop: `${theme.spacing(1)}px`,
-  paddingBottom: `${theme.spacing(1)}px`
+  paddingTop: `1px`,
+  paddingBottom: `1px`
 }))